+++ lib/white_bread/context/setup.ex
@@ -1,13 +1,6 @@
 defmodule WhiteBread.Context.Setup do
-  alias WhiteBread.Context.StepExecutor
-
   def before do
     quote do
-      def execute_step(step, state) do
-        get_steps
-          |> StepExecutor.execute_step(step, state)
-      end
-
       def get_steps do
         @sub_context_modules
          |> Enum.map(fn(sub_module) -> apply(sub_module, :get_steps, []) end)